requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
pillow>=8.0.0

# Optional dependencies for enhanced functionality
opencv-python>=4.5.0  # For advanced image processing
kenlm>=0.0.0  # For language model integration
pyctcdecode>=0.4.0  # For CTC decoding with language models

# Training and evaluation utilities
tqdm>=4.60.0
matplotlib>=3.3.0
tensorboard>=2.5.0

# Data handling
h5py>=3.1.0
scipy>=1.7.0
